src/main/java/kitchenpos/domain/MenuGroup.java
package kitchenpos.domain;

import java.util.Objects;
import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.Table;
import java.util.UUID;

@Table(name = "menu_group")
@Entity
public class MenuGroup {
    @Column(name = "id", columnDefinition = "varbinary(16)")
    @Id
    private UUID id;

    @Column(name = "name", nullable = false)
    private String name;

    public MenuGroup() {
    }

    // id를 여기서 생성하는 것과 DTO에서 toEntity할 때 생성하는 것 차이 확인?
    public MenuGroup(String name) {
        verify(name);
        this.id = UUID.randomUUID();
        this.name = name;
    }

    public UUID getId() {
        return id;
    }

    public String getName() {
        return name;
    }

    private void verify(String name) {
        if (Objects.isNull(name) || name.isEmpty()) {
            throw new IllegalArgumentException("메뉴 그룹 이름이 잘못됐어요");
        }
    }

    public void updateName(String name) {
        verify(name);
        this.name = name;
    }
}
